Cobining Basic nested loop kernel tests.
