+++ test/functional/kernel/nested-loop/CMakeLists.txt
@@ -7,31 +7,23 @@
 
 set(NESTED_LOOPTYPES Basic MultiLambda MultiLambdaParam)
 
+set( USE_RESOURCE "-resource-" "-" )
+
 #
 # Generate kernel region basic tests for each enabled RAJA back-end.
 #
 foreach( NESTED_LOOP_BACKEND ${KERNEL_BACKENDS} )
-  foreach( NESTED_LOOP_TYPE ${NESTED_LOOPTYPES} )
+  foreach( RESOURCE ${USE_RESOURCE} )
+    foreach( NESTED_LOOP_TYPE ${NESTED_LOOPTYPES} )
+      configure_file( test-kernel-nested-loop.cpp.in
+        test-kernel${RESOURCE}nested-loop-${NESTED_LOOP_TYPE}-${NESTED_LOOP_BACKEND}.cpp )
+      raja_add_test( NAME test-kernel${RESOURCE}nested-loop-${NESTED_LOOP_TYPE}-${NESTED_LOOP_BACKEND}
+        SOURCES ${CMAKE_CURRENT_BINARY_DIR}/test-kernel${RESOURCE}nested-loop-${NESTED_LOOP_TYPE}-${NESTED_LOOP_BACKEND}.cpp )
 
-    configure_file( test-kernel-nested-loop.cpp.in
-                    test-kernel-nested-loop-${NESTED_LOOP_TYPE}-${NESTED_LOOP_BACKEND}.cpp )
-    raja_add_test( NAME test-kernel-nested-loop-${NESTED_LOOP_TYPE}-${NESTED_LOOP_BACKEND}
-                   SOURCES ${CMAKE_CURRENT_BINARY_DIR}/test-kernel-nested-loop-${NESTED_LOOP_TYPE}-${NESTED_LOOP_BACKEND}.cpp )
+      target_include_directories(test-kernel${RESOURCE}nested-loop-${NESTED_LOOP_TYPE}-${NESTED_LOOP_BACKEND}.exe
+                                 PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/tests)
 
-    target_include_directories(test-kernel-nested-loop-${NESTED_LOOP_TYPE}-${NESTED_LOOP_BACKEND}.exe
-                               PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/tests)
-  endforeach()
-
-  # Resource Tests.
-  foreach( NESTED_LOOP_TYPE ${NESTED_LOOPTYPES} )
-    configure_file( test-kernel-nested-loop.cpp.in
-                    test-kernel-resource-nested-loop-${NESTED_LOOP_TYPE}-${NESTED_LOOP_BACKEND}.cpp )
-    raja_add_test( NAME test-kernel-resource-nested-loop-${NESTED_LOOP_TYPE}-${NESTED_LOOP_BACKEND}
-                   SOURCES ${CMAKE_CURRENT_BINARY_DIR}/test-kernel-resource-nested-loop-${NESTED_LOOP_TYPE}-${NESTED_LOOP_BACKEND}.cpp )
-
-    target_include_directories(test-kernel-resource-nested-loop-${NESTED_LOOP_TYPE}-${NESTED_LOOP_BACKEND}.exe
-                               PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/tests)
-
+    endforeach()
   endforeach()
 endforeach()
