+++ test/functional/kernel/nested-loop/tests/test-kernel-resource-nested-loop-MultiLambdaParam.hpp
@@ -36,9 +36,9 @@ void KernelNestedLoopTest(){
   WORKING_RES work_res{WORKING_RES::get_default()};
 
   // Allocate Tests Data
-  double* work_arrA = work_res.allocate<double>(N*N);
-  double* work_arrB = work_res.allocate<double>(N*N);
-  double* work_arrC = work_res.allocate<double>(N*N);
+  double* work_arrA = work_res.template allocate<double>(N*N);
+  double* work_arrB = work_res.template allocate<double>(N*N);
+  double* work_arrC = work_res.template allocate<double>(N*N);
 
   double* test_arrA = host_res.allocate<double>(N*N);
   double* test_arrB = host_res.allocate<double>(N*N);
